requirements.txt
autoenv==1.0.0
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.3
Werkzeug==1.0.1
WTForms==2.3.3
